# src/main/resources/public/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Documentation Tools API Voltz",
    "version": "v1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Tools Toltz"
    }
  ],
  "paths": {
    "/tools": {
      "get": {
        "tags": [
          "Tools"
        ],
        "summary": "Finds tools by Id",
        "description": "Return a list of tools",
        "operationId": "listTools",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Tools"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Tools"
        ],
        "summary": "Add a new tool",
        "operationId": "persistTools",
        "requestBody": {
          "description": "Tools object that needs to be added",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToolsNoId"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "created",
            "content": {
              "application/json:": {
                "schema": {
                  "$ref": "#/components/schemas/Tools"
                }
              }
            }
          },
          "400": {
            "description": "The max number of tag is eight (8)",
            "content": {}
          },
          "404": {
            "description": "The 'example' tag not found",
            "content": {}
          }
        }
      }
    },
    "/tools/{id}": {
      "delete": {
        "tags": [
          "Tools"
        ],
        "summary": "Delete tools by Id",
        "operationId": "removeTools",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/tag": {
      "get": {
        "tags": [
          "Tag"
        ],
        "summary": "Finds tags by Id",
        "description": "Return a list of tags",
        "operationId": "listTag",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Tag"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Tag"
        ],
        "summary": "Add a new tag",
        "operationId": "persistTag",
        "requestBody": {
          "description": "Tag object that needs to be added",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Tag"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "created",
            "content": {
              "application/json:": {
                "schema": {
                  "$ref": "#/components/schemas/Tag"
                }
              }
            }
          }
        }
      }
    },
    "/tag/{name}": {
      "delete": {
        "tags": [
          "Tag"
        ],
        "summary": "Delete tag by name",
        "operationId": "removeTag",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ToolsNoId": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "link": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "Tools": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "link": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "Tag": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        }
      }
    }
  }
}